score et acceleration vitesse

+++ src/thread/BallsBounding.java
@@ -30,7 +30,7 @@ public class BallsBounding extends CancelableThread {
 		}
 	    }
 	    try {
-		Thread.sleep(CancelableThread.TIME_TO_WAIT);
+		Thread.sleep((long) CancelableThread.TIME_TO_WAIT);
 	    } catch (InterruptedException ie) {
 		ie.printStackTrace();
 	    }
