+++ src/thread/BallsMove.java
@@ -28,7 +28,7 @@ public class BallsMove extends CancelableThread {
 		}
 	    }
 	    try {
-		Thread.sleep(CancelableThread.TIME_TO_WAIT);
+		Thread.sleep((long) CancelableThread.TIME_TO_WAIT);
 	    } catch (InterruptedException ie) {
 		ie.printStackTrace();
 	    }
